requirements.txt
aniso8601==8.0.0
attrs==19.3.0
click==7.1.2
eggit==0.0.22
flasgger==0.9.4
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
from==0.1.0
import==0.0.1
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
JsonForm==0.0.2
jsonschema==3.2.0
JsonSir==0.0.2
MarkupSafe==1.1.1
mistune==0.8.4
passlib==1.7.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.16.0
python-dateutil==2.8.1
Python-EasyConfig==0.1.7
pytz==2020.1
PyYAML==5.3.1
Resource==0.2.1
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
